backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
redis
python-dotenv
pydantic-settings
alembic
ruff